requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
astor==0.8.1
asttokens==3.0.0
attrs==25.4.0
blake3==1.0.8
cachetools==6.2.1
cbor2==5.7.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.2
coloredlogs==15.0.1
comm==0.2.3
compressed-tensors==0.11.0
contourpy==1.3.3
cuda-toolkit==13.0.1
cupy-cuda12x==13.6.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
einops==0.8.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.121.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastrlock==0.8.3
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.9.0
gguf==0.17.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
inquirerpy==0.3.4
interegular==0.3.3
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.2.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.8.5
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime==13.0.88
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnx==1.19.1
onnx-ir==0.1.12
onnx_graphsurgeon==0.5.8
onnxruntime-gpu==1.23.2
onnxscript==0.5.6
openai==2.7.1
openai-harmony==0.0.8
opencv-python-headless==4.12.0.88
outlines_core==0.2.11
packaging==25.0
parso==0.8.5
partial-json-parser==0.2.1.1.post6
pexpect==4.9.0
pfzy==0.3.4
pillow==12.0.0
platformdirs==4.4.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.23
pycuda==2025.1.2
pydantic==2.12.4
pydantic-extra-types==2.10.6
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytools==2025.2.5
PyYAML==6.0.2
pyzmq==27.1.0
ray==2.51.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rpds-py==0.28.0
safetensors==0.6.2
scipy==1.16.3
sentencepiece==0.2.1
sentry-sdk==2.43.0
setproctitle==1.3.7
setuptools==79.0.1
shellingham==1.5.4
siphash24==1.8
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
tensorrt==10.9.0.34
tensorrt_cu12==10.9.0.34
tensorrt_cu12_bindings==10.9.0.34
tensorrt_cu12_libs==10.9.0.34
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.4.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vllm==0.11.0
watchfiles==1.1.1
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.14
xformers==0.0.32.post1
xgrammar==0.1.25
yarl==1.22.0
